==> src/0_run_simulation.R <==

library(here)
source(here("1_simulate_data.R"))
source(here("2_calculate_truth.R"))
source(here("3_run_simulation_for_point_estimators.R"))
source(here("4_calc_performance.R"))
source(here("5_run_bootstrap_variance.R"))
source(here("6_calc_bootstrap_performance.R"))

source(here("7_simulate_data_null.R"))
source(here("8_run_null_simulations_for_point_estimates.R"))
source(here("9_calc_performance_null.R"))

source(here("10_create_latex_tables.R"))
source(here("11_example_DAG.R"))
